Refactor code by removing unused imports, adding missing type checks, and enabling noUnusedLocals in tsconfig

+++ src/Utils/chat-utils.ts
@@ -387,7 +387,7 @@ export const extractSyncdPatches = async (result: BinaryNode, options: RequestIn
 						content = Buffer.from(Object.values(content))
 					}
 
-					const syncd = proto.SyncdPatch.decode(content as Uint8Array)
+					const syncd = proto.SyncdPatch.decode(content)
 					if (!syncd.version) {
 						syncd.version = { version: +collectionNode.attrs.version! + 1 }
 					}
@@ -725,7 +725,7 @@ export const chatModificationToAppPatch = (mod: ChatModification, jid: string) =
 			apiVersion: 1,
 			operation: OP.SET
 		}
-	}  else if ('addLabel' in mod) {
+	} else if ('addLabel' in mod) {
 		patch = {
 			syncAction: {
 				labelEditAction: {
